pyproject.toml
[project]
name = "airflow-training"
version = "0.1.0"
description = ""
authors = [
    {name = "Jiliar Silgado",email = "jiliar.silgado@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "apache-airflow (>=3.0.1,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "apache-airflow-providers-postgres (>=6.2.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
